Warning fixes for cryptobox 0.2.0

+++ Cryptobox/Cryptobox/CBVector+Internal.h
@@ -15,6 +15,6 @@
 
 - (nonnull uint8_t *)dataArray;
 
-- (uint32_t)length;
+- (size_t)length;
 
 @end
